requirements.txt
defusedxml==0.7.1
exceptiongroup==1.1.2
fonttools==4.42.0
fpdf2==2.7.5
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
pytest==7.4.0
tomli==2.0.1
